Implement basic generic analysis
This completes the implementation of generic analysis (modulo bugs
and unexpectedly missing features). The code now parses the signatures
of pairs of classes and compares the generic type parameters for α-equivalence.
It doesn't attempt to do any more complex analysis than this, and it also
treats all generic type changes as purely source-incompatible changes. This
is not strictly correct with regards to JLS 9 §13.4.5, which states:

> Changing the first bound of a type parameter of a class may change
> the erasure (§4.6) of any member that uses that type parameter in
> its own type, and this may affect binary compatibility. The change
> of such a bound is analogous to the change of the first bound of a
> type parameter of a method or constructor (§13.4.13).

More detailed analysis of the initial bounds may be necessary.

Fix #2